Rework after peer review: deleted file, renamed other file, move implementation of SingletonFreshName to util.py
