+++ src/nagini_translation/lib/util.py
@@ -373,3 +373,24 @@ class OldExpressionTransformer(ast.NodeTransformer):
                             args=[ast.Num(n=index)],
                             keywords=[])
         return node
+
+
+class SingletonFreshName:
+    """
+    This class wraps the fresh name facility in scope by using it only when
+    a new name is given. It is designed to store and retrieve the fresh name
+    based on the original name given.
+    """
+
+    def __init__(self, scope: 'PythonScope') -> None:
+        self._fresh_name_dict = {}
+        self._scope = scope
+
+    def __call__(self, name: str) -> str:
+        """
+        Returns a fresh name for a given name and scope if enquired for the
+        first time, otherwise returns the previously given fresh name.
+        """
+        if name not in self._fresh_name_dict:
+            self._fresh_name_dict[name] = self._scope.get_fresh_name(name)
+        return self._fresh_name_dict[name]
